Add files via upload
In this project, I navigated the intricacies of date cleaning, preparation, and the construction of a professional and automated dashboard. Building on array capabilities, I explored dynamic lookup functions like INDIRECT, ADDRESS, OFFSET, and INDEX. The project culminated in the setup of a dashboard, equipping me with skills to visualize data using graphs and charts, construct data models, and incorporate interactivity for enhanced data presentation.